+++ src/main.cpp
@@ -240,6 +240,9 @@ void waitForPowerRelease() {
 void enterDeepSleep() {
   HalPowerManager::Lock powerLock;  // Ensure we are at normal CPU frequency for sleep preparation
   APP_STATE.lastSleepFromReader = activityManager.isReaderActivity();
+  if (APP_STATE.lastSleepFromReader) {
+    READING_STATS.checkpointActiveSession();
+  }
   APP_STATE.saveToFile();
 
   activityManager.goToSleep();